modify Service.

+++ src/main/java/com/example/demo/repository/VehicleRepository.java
@@ -19,7 +19,7 @@ public class VehicleRepository {
     @Resource
     MysqlVehicleRepository mysqlVehicleRepository;
 
-    public Integer save(VehicleInformationPO vehicleInformationPO){
+    public Long save(VehicleInformationPO vehicleInformationPO){
         return mysqlVehicleRepository.save(vehicleInformationPO);
     }
